chore: split signal adapter from player index
